static analysis
